+++ CHANGELOG.md
@@ -4,6 +4,7 @@ LifterLMS Changelog
 v3.16.7 - 2018-02-??
 --------------------
 
++ Added dashboard links to the WP Admin Bar to allow existing the course builder to various areas of the dashboard
 + Added data attribute to progress bars so JS (or CSS) can read the progress of a progress bar. Thanks [@dineshchouhan](https://github.com/dineshchouhan)!
 + Fixed php `max_input_vars` issue causing a 400 Bad Request error when trying to save large courses in the course builder